week_04/Assignment_12/requirements.txt
langchain-openai>=0.1.0
pillow>=10.0.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv